refactor: Board 인터페이스화
